+++ src/main/java/com/tilialog/App.java
@@ -1,31 +1,65 @@
 package com.tilialog;
 
+import com.github.zafarkhaja.semver.Version;
+import com.jcabi.xml.XMLDocument;
 import com.tilialog.ui.MainFrame;
 import java.io.File;
+import java.io.IOException;
 import javax.swing.SwingUtilities;
 
 public class App implements Runnable {
-    private MainFrame mainFrame;
-    private Settings settings;
 
-    public App(Settings settings) {
-        this.settings = settings;
+    private MainFrame mainFrame;
+
+    public App(MainFrame mainFrame) {
+        this.mainFrame = mainFrame;
     }
 
     @Override
     public void run() {
-        mainFrame = new MainFrame(this.settings);
+        mainFrame.setVisible();
     }
 
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(
-            new App(
-                new TlSettings(
-                    new TlTextFile(
-                        new File("settings.list")
-                    )
+    public static void main(String[] args) throws IOException {
+        AppIsOutdated appIsOutdated = new CachedAppIsOutdated(
+            new TlAppIsOutdated(
+                Version.valueOf(
+                    new XMLDocument(
+                        new TlTextFile(
+                            new File("pom.xml")
+                        ).readAll()
+                    ).xpath("/*[local-name()=\"project\"]/*[local-name()=\"version\"]/text()").get(0)
                 )
             )
         );
+
+        MainFrame mainFrame = new MainFrame(
+            new TlSettings(
+                new TlTextFile(
+                    new File("settings.list")
+                )
+            ),
+            appIsOutdated
+        );
+
+        SwingUtilities.invokeLater(
+            new App(
+                mainFrame
+            )
+        );
+
+        new Thread(new Runnable() {
+            @Override
+            public void run() {
+                if (appIsOutdated.asBoolean()) {
+                    SwingUtilities.invokeLater(new Runnable() {
+                        @Override
+                        public void run() {
+                            mainFrame.showUpdatesMessageDialog();
+                        }
+                    });
+                }
+            }
+        }).start();
     }
 }