+++ src/main/java/com/tilialog/AppIsOutdated.java
@@ -1,0 +1,10 @@
+package com.tilialog;
+
+import com.github.zafarkhaja.semver.Version;
+
+public interface AppIsOutdated {
+
+    public boolean asBoolean();
+
+    public Version runningVersion();
+}