+++ src/main/java/com/tilialog/ui/AboutPanel.java
@@ -1,15 +1,18 @@
 package com.tilialog.ui;
 
+import java.net.URI;
 import javax.swing.JEditorPane;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 import java.awt.Desktop;
+import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-import java.awt.event.MouseListener;
 import java.io.IOException;
 
 public class AboutPanel {
+
     private final String siteUrl = "https://github.com/linden2015/tilialog";
+
     private JPanel panel;
 
     public AboutPanel() {
@@ -30,34 +33,19 @@ public class AboutPanel {
         panel.add(urlLabel);
     }
 
-    private class UrlLabelMouseListener implements MouseListener {
-        @Override
-        public void mouseClicked(MouseEvent e) {
-        }
+    private class UrlLabelMouseListener extends MouseAdapter {
 
         @Override
         public void mousePressed(MouseEvent me) {
             try {
                 Desktop.getDesktop().browse(
-                    java.net.URI.create(siteUrl)
+                    URI.create(siteUrl)
                 );
             } catch (IOException e) {
                 throw new RuntimeException(
-                    "Unrecoverable error while opening project site.", e
+                    "Unrecoverable error while opening project webpage.", e
                 );
             }
-        }
-
-        @Override
-        public void mouseReleased(MouseEvent e) {
-        }
-
-        @Override
-        public void mouseEntered(MouseEvent e) {
-        }
-
-        @Override
-        public void mouseExited(MouseEvent e) {
         }
     }
 }